backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
pandas==2.3.2
openpyxl==3.1.5
python-dotenv==1.1.1
django-cors-headers==4.9.0
openai==1.108.1
requests==2.32.5